+++ sphinxcontrib/needs/layout.py
@@ -155,7 +155,9 @@ class LayoutHandler:
         else:
             self.fromdocname = fromdocname
 
+        # For ReadTheDocs Theme we need to add 'rtd-exclude-wy-table'.
         classes = ["need", "needs_grid_" + self.layout["grid"], "needs_layout_" + self.layout_name]
+        classes.extend(app.config.needs_table_classes)
 
         self.style = style or self.need["style"] or getattr(self.app.config, "needs_default_style", None)